@@ -1,11 +1,11 @@
 {
   "name": "client",
-  "version": "0.1.0",
+  "version": "3.0.0",
   "private": true,
   "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^6.4.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.17.0",
     "react-scripts": "5.0.1"
   },
   "scripts": {
